set initial theme on app mount
Since the themeSelector isn't mounted until the app menu is open, the theme initialization doesn't work right away.

This extracts the theme logic to a separate composable `useTheming` and then calls `useTheming`  when the app is mounted (in addition to within the theme selector component).
